# internal/api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Enduro API
    version: ""
host: localhost:9000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /storage/{aip_id}/download:
        get:
            tags:
                - storage
            summary: download storage
            description: Download package by AIPID
            operationId: storage#download
            parameters:
                - name: aip_id
                  in: path
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
                        format: byte
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/StorageDownloadNotFoundResponseBody'
                        required:
                            - message
                            - aip_id
            schemes:
                - http
    /storage/{aip_id}/submit:
        post:
            tags:
                - storage
            summary: submit storage
            description: Start the submission of a package
            operationId: storage#submit
            parameters:
                - name: aip_id
                  in: path
                  required: true
                  type: string
                - name: SubmitRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/StorageSubmitRequestBody'
                    required:
                        - name
            responses:
                "202":
                    description: Accepted response.
                    schema:
                        $ref: '#/definitions/StorageSubmitResponseBody'
                        required:
                            - url
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/StorageSubmitNotValidResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/StorageSubmitNotAvailableResponseBody'
            schemes:
                - http
    /storage/{aip_id}/update:
        post:
            tags:
                - storage
            summary: update storage
            description: Signal the storage service that an upload is complete
            operationId: storage#update
            parameters:
                - name: aip_id
                  in: path
                  required: true
                  type: string
            responses:
                "202":
                    description: Accepted response.
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/StorageUpdateNotValidResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/StorageUpdateNotAvailableResponseBody'
            schemes:
                - http
...
definitions:
...
    StorageDownloadNotFoundResponseBody:
        title: StorageDownloadNotFoundResponseBody
        type: object
        properties:
            aip_id:
                type: string
                description: Identifier of missing package
                example: Distinctio ut.
            message:
                type: string
                description: Message of error
                example: Esse nulla aliquid facere qui tenetur corrupti.
        description: Storage package not found
        example:
            aip_id: Sunt totam.
            message: Facilis ex praesentium.
        required:
            - message
            - aip_id
...
    StorageSubmitNotAvailableResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: submit_not_available_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    StorageSubmitNotValidResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: submit_not_valid_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    StorageSubmitRequestBody:
        title: StorageSubmitRequestBody
        type: object
        properties:
            name:
                type: string
                example: Officia maxime.
        example:
            name: Quos quo in ab.
        required:
            - name
    StorageSubmitResponseBody:
        title: StorageSubmitResponseBody
        type: object
        properties:
            url:
                type: string
                example: Quod qui.
        example:
            url: Dolore nam.
        required:
            - url
    StorageUpdateNotAvailableResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: update_not_available_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    StorageUpdateNotValidResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: update_not_valid_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
...